Refactor code to use SOLID principles
